+++ zmscitizenapi/src/Zmscitizenapi/BaseController.php
@@ -5,10 +5,6 @@ namespace BO\Zmscitizenapi;
 use \Psr\Http\Message\RequestInterface;
 use \Psr\Http\Message\ResponseInterface;
 
-/**
- * @SuppressWarnings(NumberOfChildren)
- *
- */
 abstract class BaseController extends \BO\Slim\Controller
 {
     public function __invoke(RequestInterface $request, ResponseInterface $response, array $args)
@@ -105,120 +101,6 @@ abstract class BaseController extends \BO\Slim\Controller
     {
         return array_search($value, $self) === $index;
     }
-
-
-    // ----------------- ZmsApiClient ---------------------
-
-    protected function getSources()
-    {
-        $requestUrl = '/source/dldb/?resolveReferences=2';
-
-        return \App::$http->readGetResult($requestUrl)->getEntity();
-    }
-
-    protected function getScopes()
-    {
-        $sources = $this->getSources();
-        return $sources->scopes;
-    }
-
-    protected function getFreeDays($providers, $requests, $firstDay, $lastDay)
-    {
-        $requestUrl = '/calendar/';
-        $dataPayload = [
-            'firstDay' => $firstDay,
-            'lastDay' => $lastDay,
-            'providers' => $providers,
-            'requests' => $requests,
-        ];
-
-        return \App::$http->readPostResult($requestUrl, $dataPayload)->getEntity();
-    }
-
-    protected function getFreeTimeslots($providers, $requests, $firstDay, $lastDay)
-    {
-        $requestUrl = '/process/status/free/';
-        $dataPayload = [
-            'firstDay' => $firstDay,
-            'lastDay' => $lastDay,
-            'providers' => $providers,
-            'requests' => $requests,
-        ];
-
-        return \App::$http->readPostResult($requestUrl, $dataPayload)->getEntity();
-    }
-
-    protected function reserveTimeslot($appointmentProcess, $serviceIds, $serviceCounts)
-    {
-        $requests = [];
-
-        foreach ($serviceIds as $index => $serviceId) {
-            $count = intval($serviceCounts[$index]);
-            for ($i = 0; $i < $count; $i++) {
-                $requests[] = [
-                    'id' => $serviceId,
-                    'source' => 'dldb'
-                ];
-            }
-        }
-
-        $appointmentProcess['requests'] = $requests;
-
-        return \App::$http->readPostResult('/process/status/reserved/', $appointmentProcess)->getEntity();
-    }
-
-    protected function submitClientData($process)
-    {
-        $url = "/process/{$process['id']}/{$process['authKey']}/";
-        return \App::$http->readPostResult($url, $process)->getEntity();
-    }
-
-    protected function preconfirmProcess($process)
-    {
-        $url = '/process/status/preconfirmed/';
-        return \App::$http->readPostResult($url, $process)->getEntity();
-    }
-
-    protected function confirmProcess($process)
-    {
-        $url = '/process/status/confirmed/';
-        return \App::$http->readPostResult($url, $process)->getEntity();
-    }
-
-    protected function cancelAppointment($process)
-    {
-        $url = "/process/{$process['id']}/{$process['authKey']}/";
-        return \App::$http->readDeleteResult($url, $process)->getEntity();
-    }
-
-    protected function sendConfirmationEmail($process)
-    {
-        $url = "/process/{$process['id']}/{$process['authKey']}/confirmation/mail/";
-        return \App::$http->readPostResult($url, $process)->getEntity();
-    }
-
-    protected function sendPreconfirmationEmail($process)
-    {
-        $url = "/process/{$process['id']}/{$process['authKey']}/preconfirmation/mail/";
-        return \App::$http->readPostResult($url, $process)->getEntity();
-    }
-
-    protected function sendCancelationEmail($process)
-    {
-        $url = "/process/{$process['id']}/{$process['authKey']}/delete/mail/";
-        return \App::$http->readPostResult($url, $process)->getEntity();
-    }
-
-
-    protected function getProcessById($processId, $authKey)
-    {
-        $resolveReferences = 2;
-        $process = \App::$http->readGetResult("/process/{$processId}/{$authKey}/", [
-            'resolveReferences' => $resolveReferences
-        ])->getEntity();
-
-        return $process;
-    }
-
+    
 
 }
