+++ zmscitizenapi/src/Zmscitizenapi/Controllers/AppointmentUpdate.php
@@ -1,12 +1,13 @@
 <?php
 
-namespace BO\Zmscitizenapi;
+namespace BO\Zmscitizenapi\Controllers;
 
+use BO\Zmscitizenapi\BaseController;
 use BO\Slim\Render;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 
-class AppointmentCancel extends BaseController
+class AppointmentUpdate extends BaseController
 {
     public function readResponse(RequestInterface $request, ResponseInterface $response, array $args)
     {
