+++ zmscitizenapi/src/Zmscitizenapi/Controllers/AvailableDaysList.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace BO\Zmscitizenapi\Controllers;
+
+use BO\Zmscitizenapi\BaseController;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
+use BO\Zmscitizenapi\Services\AvailableDaysService;
+
+class AvailableDaysList extends BaseController
+{
+    protected $availableDaysService;
+
+    public function __construct()
+    {
+        $this->availableDaysService = new AvailableDaysService();
+    }
+
+    public function readResponse(RequestInterface $request, ResponseInterface $response, array $args)
+    {
+        $queryParams = $request->getQueryParams();
+
+        $result = $this->availableDaysService->getAvailableDays($queryParams);
+        
+        return $this->createJsonResponse($response, $result, $result['status']);
+    }
+
+    private function createJsonResponse(ResponseInterface $response, array $content, int $statusCode): ResponseInterface
+    {
+        $response = $response->withStatus($statusCode)
+                             ->withHeader('Content-Type', 'application/json');
+        $response->getBody()->write(json_encode($content, JSON_NUMERIC_CHECK));
+        return $response;
+    }
+}
