+++ zmscitizenapi/src/Zmscitizenapi/Controllers/CaptchaGet.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace BO\Zmscitizenapi\Controllers;
+
+use BO\Zmscitizenapi\BaseController;
+use BO\Slim\Render;
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\ResponseInterface;
+use BO\Zmscitizenapi\Services\CaptchaService;
+
+class CaptchaGet extends BaseController
+{
+    protected $captchaService;
+
+    public function __construct()
+    {
+        // Initialize the CaptchaService
+        $this->captchaService = new CaptchaService();
+    }
+
+    public function readResponse(RequestInterface $request, ResponseInterface $response, array $args)
+    {
+        $captchaDetails = $this->captchaService->getCaptchaDetails();
+
+        return Render::withJson($response, $captchaDetails);
+    }
+}
