+++ zmscitizenapi/src/Zmscitizenapi/Services/AvailableAppointmentsService.php
@@ -1,0 +1,152 @@
+<?php
+
+namespace BO\Zmscitizenapi\Services;
+
+use BO\Zmsentities\Calendar as CalendarEntity;
+
+class AvailableAppointmentsService
+{
+    public function getAvailableAppointments(array $queryParams)
+    {
+        $date = $queryParams['date'] ?? null;
+        $officeId = $queryParams['officeId'] ?? null;
+        $serviceIds = isset($queryParams['serviceId']) ? explode(',', $queryParams['serviceId']) : [];
+        $serviceCounts = isset($queryParams['serviceCount']) ? explode(',', $queryParams['serviceCount']) : [];
+
+        $errors = $this->validateQueryParams($date, $officeId, $serviceIds, $serviceCounts);
+
+        if (!empty($errors)) {
+            return ['errors' => $errors, 'status' => 400];
+        }
+
+        try {
+            $calendar = new CalendarEntity();
+            $calendar->firstDay = $this->getInternalDateFromISO($date);
+            $calendar->lastDay = $this->getInternalDateFromISO($date);
+            $calendar->providers = [['id' => $officeId, 'source' => 'dldb']];
+
+            $calendar->requests = [];
+            foreach ($serviceIds as $index => $serviceId) {
+                $slotCount = isset($serviceCounts[$index]) ? intval($serviceCounts[$index]) : 1;
+                for ($i = 0; $i < $slotCount; $i++) {
+                    $calendar->requests[] = [
+                        'id' => $serviceId,
+                        'source' => 'dldb',
+                        'slotCount' => 1,
+                    ];
+                }
+            }
+
+            $freeSlots = \App::$http->readPostResult('/process/status/free/', $calendar);
+            if (!$freeSlots || !method_exists($freeSlots, 'getCollection')) {
+                throw new \Exception('Invalid response from API');
+            }
+
+            return $this->processFreeSlots($freeSlots->getCollection());
+
+        } catch (\Exception $e) {
+            error_log('Error in AvailableAppointmentsService: ' . $e->getMessage());
+            return [
+                'appointmentTimestamps' => [],
+                'errorCode' => 'internalError',
+                'errorMessage' => 'An error occurred while fetching available appointments',
+                'status' => 500,
+            ];
+        }
+    }
+
+    private function validateQueryParams($date, $officeId, $serviceIds, $serviceCounts)
+    {
+        $errors = [];
+    
+        if (!$date) {
+            $errors[] = [
+                'type' => 'field',
+                'msg' => 'date is required and must be a valid date',
+                'path' => 'date',
+                'location' => 'body'
+            ];
+        }
+    
+        if (!$officeId || !is_numeric($officeId)) {
+            $errors[] = [
+                'type' => 'field',
+                'msg' => 'officeId should be a 32-bit integer',
+                'path' => 'officeId',
+                'location' => 'body'
+            ];
+        }
+    
+        if (empty($serviceIds[0]) || !preg_match('/^\d+(,\d+)*$/', implode(',', $serviceIds))) {
+            $errors[] = [
+                'type' => 'field',
+                'msg' => 'serviceId should be a comma-separated string of integers',
+                'path' => 'serviceId',
+                'location' => 'body'
+            ];
+        }
+    
+        if (empty($serviceCounts[0]) || !preg_match('/^\d+(,\d+)*$/', implode(',', $serviceCounts))) {
+            $errors[] = [
+                'type' => 'field',
+                'msg' => 'serviceCount should be a comma-separated string of integers',
+                'path' => 'serviceCount',
+                'location' => 'body'
+            ];
+        }
+    
+        return $errors;
+    }
+    
+
+    private function processFreeSlots($freeSlots)
+    {
+        if (empty($freeSlots)) {
+            return [
+                'appointmentTimestamps' => [],
+                'errorCode' => 'appointmentNotAvailable',
+                'errorMessage' => 'Der von Ihnen gewählte Termin ist leider nicht mehr verfügbar',
+                'status' => 404,
+            ];
+        }
+
+        $currentTimestamp = time();
+        $appointmentTimestamps = [];
+
+        foreach ($freeSlots as $slot) {
+            foreach ($slot->appointments as $appointment) {
+                $timestamp = (int)$appointment->date;
+                if (!in_array($timestamp, $appointmentTimestamps) && $timestamp > $currentTimestamp) {
+                    $appointmentTimestamps[] = $timestamp;
+                }
+            }
+        }
+
+        if (empty($appointmentTimestamps)) {
+            return [
+                'appointmentTimestamps' => [],
+                'errorCode' => 'appointmentNotAvailable',
+                'errorMessage' => 'Der von Ihnen gewählte Termin ist leider nicht mehr verfügbar',
+                'status' => 404,
+            ];
+        }
+
+        sort($appointmentTimestamps);
+
+        return [
+            'appointmentTimestamps' => $appointmentTimestamps,
+            'lastModified' => round(microtime(true) * 1000),
+            'status' => 200,
+        ];
+    }
+
+    private function getInternalDateFromISO($dateString)
+    {
+        $date = new \DateTime($dateString);
+        return [
+            'day' => (int) $date->format('d'),
+            'month' => (int) $date->format('m'),
+            'year' => (int) $date->format('Y'),
+        ];
+    }
+}
