+++ zmscitizenapi/tests/Zmscitizenapi/AppointmentCancelTest.php
@@ -5,7 +5,7 @@ namespace BO\Zmscitizenapi\Tests;
 class AppointmentCancelTest extends Base
 {
 
-    protected $classname = "AppointmentCancel";
+    protected $classname = "\BO\Zmscitizenapi\Controllers\AppointmentCancel";
 
     public function testRendering() {
         $responseData = $this->renderJson(method: 'POST');
