+++ zmscitizenapi/tests/Zmscitizenapi/AppointmentPreconfirmTest.php
@@ -5,7 +5,7 @@ namespace BO\Zmscitizenapi\Tests;
 class AppointmentPreconfirmTest extends Base
 {
 
-    protected $classname = "AppointmentPreconfirm";
+    protected $classname = "\BO\Zmscitizenapi\Controllers\AppointmentPreconfirm";
 
     public function testRendering() {
         $responseData = $this->renderJson(method: 'POST');
