+++ zmscitizenapi/tests/Zmscitizenapi/AppointmentReserveTest.php
@@ -5,7 +5,7 @@ namespace BO\Zmscitizenapi\Tests;
 class AppointmentReserveTest extends Base
 {
 
-    protected $classname = "AppointmentReserve";
+    protected $classname = "\BO\Zmscitizenapi\Controllers\AppointmentReserve";
 
     public function testRendering() {
         $responseData = $this->renderJson(method: 'POST');
