+++ zmscitizenapi/tests/Zmscitizenapi/AppointmentUpdateTest.php
@@ -5,7 +5,7 @@ namespace BO\Zmscitizenapi\Tests;
 class AppointmentUpdateTest extends Base
 {
 
-    protected $classname = "AppointmentUpdate";
+    protected $classname = "\BO\Zmscitizenapi\Controllers\AppointmentUpdate";
 
     public function testRendering() {
         $responseData = $this->renderJson(method: 'POST');
