+++ zmscitizenapi/tests/Zmscitizenapi/ScopesListTest.php
@@ -5,7 +5,7 @@ namespace BO\Zmscitizenapi\Tests;
 class ScopesListTest extends Base
 {
 
-    protected $classname = "ScopesList";
+    protected $classname = "\BO\Zmscitizenapi\Controllers\ScopesList";
 
     public function testRendering() {
         $this->setApiCalls([
